requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Telethon==1.37.0
aiohttp==3.9.1
gunicorn==21.2.0
Werkzeug==3.0.0
cryptography==42.0.0
python-dotenv==1.0.0
